chore: rename mock_input_data module
