+++ powersimdata/tests/test_mocks.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 
 from powersimdata.tests.mock_grid import MockGrid
-from powersimdata.tests.mock_input_data import MockProfileInput
+from powersimdata.tests.mock_profile_input import MockProfileInput
 from powersimdata.tests.mock_scenario import MockScenario
 from powersimdata.tests.mock_scenario_info import MockScenarioInfo
 
@@ -130,13 +130,13 @@ class TestMockInputData:
         grid = MockGrid(grid_attrs={"plant": mock_plant})
         return grid
 
-    def test_create_mock_input_data(self, grid):
+    def test_create_mock_profile_input(self, grid):
         assert MockProfileInput(grid) is not None
 
     def test_happy_case(self, grid):
-        mock_input_data = MockProfileInput(grid, periods=3)
+        mock_profile_input = MockProfileInput(grid, periods=3)
 
-        demand = mock_input_data.get_data({}, "demand")
+        demand = mock_profile_input.get_data({}, "demand")
         expected_demand_values = np.array(
             [
                 [0.355565, 0.453391, 0.563135],
@@ -147,7 +147,7 @@ class TestMockInputData:
         expected_zone_ids = np.array([1, 2, 3])
         self._assert_profile(demand, expected_demand_values, expected_zone_ids)
 
-        wind = mock_input_data.get_data({}, "wind")
+        wind = mock_profile_input.get_data({}, "wind")
         expected_wind_values = np.array(
             [
                 [0.460527],
@@ -158,7 +158,7 @@ class TestMockInputData:
         expected_wind_plant_ids = np.array([102])
         self._assert_profile(wind, expected_wind_values, expected_wind_plant_ids)
 
-        solar = mock_input_data.get_data({}, "solar")
+        solar = mock_profile_input.get_data({}, "solar")
         expected_solar_values = np.array(
             [
                 [0.305825],
@@ -169,7 +169,7 @@ class TestMockInputData:
         expected_solar_plant_ids = np.array([101])
         self._assert_profile(solar, expected_solar_values, expected_solar_plant_ids)
 
-        hydro = mock_input_data.get_data({}, "hydro")
+        hydro = mock_profile_input.get_data({}, "hydro")
         expected_hydro_values = np.array(
             [
                 [0.577557],
@@ -181,53 +181,53 @@ class TestMockInputData:
         self._assert_profile(hydro, expected_hydro_values, expected_hydro_plant_ids)
 
     def test_multiple_get_data_calls_returns_same_data(self, grid):
-        mock_input_data = MockProfileInput(grid)
+        mock_profile_input = MockProfileInput(grid)
 
         for type in ("demand", "wind", "solar", "hydro"):
-            profile1 = mock_input_data.get_data({}, type)
-            profile2 = mock_input_data.get_data({}, type)
+            profile1 = mock_profile_input.get_data({}, type)
+            profile2 = mock_profile_input.get_data({}, type)
             pd.testing.assert_frame_equal(profile1, profile2)
 
     def test_no_start_time(self, grid):
-        mock_input_data = MockProfileInput(
+        mock_profile_input = MockProfileInput(
             grid, start_time=None, end_time="2016-01-01 02:00", periods=3, freq="H"
         )
-        demand = mock_input_data.get_data({}, "demand")
+        demand = mock_profile_input.get_data({}, "demand")
         self._assert_dates(demand.index)
 
     def test_no_end_time(self, grid):
-        mock_input_data = MockProfileInput(
+        mock_profile_input = MockProfileInput(
             grid, start_time="2016-01-01 00:00", end_time=None, periods=3, freq="H"
         )
-        demand = mock_input_data.get_data({}, "demand")
+        demand = mock_profile_input.get_data({}, "demand")
         self._assert_dates(demand.index)
 
     def test_no_period(self, grid):
-        mock_input_data = MockProfileInput(
+        mock_profile_input = MockProfileInput(
             grid,
             start_time="2016-01-01 00:00",
             end_time="2016-01-01 02:00",
             periods=None,
             freq="H",
         )
-        demand = mock_input_data.get_data({}, "demand")
+        demand = mock_profile_input.get_data({}, "demand")
         self._assert_dates(demand.index)
 
     def test_no_freq(self, grid):
-        mock_input_data = MockProfileInput(
+        mock_profile_input = MockProfileInput(
             grid,
             start_time="2016-01-01 00:00",
             end_time="2016-01-01 02:00",
             periods=3,
             freq=None,
         )
-        demand = mock_input_data.get_data({}, "demand")
+        demand = mock_profile_input.get_data({}, "demand")
         self._assert_dates(demand.index)
 
     def test_raise_if_no_profile_specified(self, grid):
         with pytest.raises(ValueError) as exc:
-            mock_input_data = MockProfileInput(grid)
-            mock_input_data.get_data({}, "fusion")
+            mock_profile_input = MockProfileInput(grid)
+            mock_profile_input.get_data({}, "fusion")
         assert "No profile specified for fusion!" in str(exc.value)
 
     def test_raise_if_all_date_range_fields_present(self, grid):
